tcl/stm32l5x|u5x: dynamically disable the work-area when RDP = 0x55
When RDP = 0x55, only a part of the RAM is set as non-secure by the
secure application using the Secure Attribution Unit (SAU).

This non-secure RAM cannot be auto-detected by the debugger,
so to avoid programming issues, by default do not use the work-area
and fall back to flash programming without flash loader (slower).

If the user knows about the current SAU partitioning, he can provide
the work-area address and size then force the usage of the loader by
setting USE_LOADER_IN_RDP_05 to 1.

Signed-off-by: Tarek BOCHKATI <tarek.bouchkati@st.com>